# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/auth/callBack": {
            "get": {
                "description": "使用code进行最终登录同时异步用来初始化这个用户,会返回一个token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "使用code进行最终登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "github重定向的code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "初始化成功!",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.CallBack"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "get": {
                "description": "github用户登录授权接口,会自动重定向到github的授权接口上",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "github用户登录授权接口",
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "get": {
                "description": "登出之后会把jwt加到黑名单里面去",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "登出",
                "responses": {
                    "200": {
                        "description": "登出成功!",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getDomain": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户 ID 获取用户的领域",
                "responses": {
                    "200": {
                        "description": "领域获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.DomainResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    },
                    "404": {
                        "description": "用户未找到",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getEvaluation": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据userid获取用户评价",
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.EvaluationResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getInfo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "从userid获取用户",
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getNation": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据用户 ID 获取用户所在国家",
                "responses": {
                    "200": {
                        "description": "国家获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.NationResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    },
                    "404": {
                        "description": "用户未找到",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getRank": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据userid获取用户的score的排行榜",
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.Ranking"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/getUserInfo": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据userid获取用户详细信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户的user_id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户信息获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/user/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "根据国家和领域搜索用户",
                "parameters": [
                    {
                        "type": "string",
                        "description": "国家，选择性参数",
                        "name": "nation",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "领域，选择性参数",
                        "name": "domain",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页参数表示这是第几页",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页返回的用户数量，建议一次返回10个",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户搜索成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.SearchResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Leaderboard": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "Bio": {
                    "description": "用户的个人简介",
                    "type": "string"
                },
                "avatar_url": {
                    "description": "用户头像的 URL",
                    "type": "string"
                },
                "blog": {
                    "description": "博客连接",
                    "type": "string"
                },
                "collaborators": {
                    "description": "协作者的数量",
                    "type": "integer"
                },
                "company": {
                    "description": "用户所属的公司",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "evaluation": {
                    "description": "评估",
                    "type": "string"
                },
                "followers": {
                    "description": "粉丝数",
                    "type": "integer"
                },
                "following": {
                    "description": "关注数",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "description": "地区",
                    "type": "string"
                },
                "login_name": {
                    "description": "用户的登录名",
                    "type": "string"
                },
                "name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "nationality": {
                    "description": "国籍",
                    "type": "string"
                },
                "public_repos": {
                    "description": "用户公开的仓库的数量",
                    "type": "integer"
                },
                "score": {
                    "description": "评分",
                    "type": "number"
                },
                "total_private_repos": {
                    "description": "用户的私有仓库总数",
                    "type": "integer"
                }
            }
        },
        "response.CallBack": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "response.DomainResp": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.Err": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
        "response.EvaluationResp": {
            "type": "object",
            "properties": {
                "evaluation": {
                    "type": "string"
                }
            }
        },
        "response.NationResp": {
            "type": "object",
            "properties": {
                "nation": {
                    "type": "string"
                }
            }
        },
        "response.Ranking": {
            "type": "object",
            "properties": {
                "leaderboard": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Leaderboard"
                    }
                }
            }
        },
        "response.SearchResp": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.User"
                    }
                }
            }
        },
        "response.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "response.User": {
            "type": "object",
            "properties": {
                "domain": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        }
    }
}